requirements.txt
keboola.component==1.3.7
keboola.utils==1.1.0
keboola.http-client==1.0.0
mock==4.0.3
freezegun==1.1.0
requests==2.27.1
